# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: invoice and bills system
  version: 1.0.0

tags:
  - name: Products

paths:
  /products:
    get:
      tags: [Products]
      summary: get products
      operationId: getProducts
      responses:
        '200':
          description: list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/productResponse'

    post:
        tags: [ Products ]
        summary: create a product
        operationId: saveProduct
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRequest'
        responses:
          '201':
            description: product created
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/productResponse'

  /products/{id}:
    get:
      tags: [Products]
      summary: get product by id
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/productResponse'
        '404':
          description: product not found

    put:
      tags: [Products]
      summary: update a product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/productResponse'

    patch:
      tags: [Products]
      summary: delete a product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: product deleted
        '404':
          description: product not found



components:
  schemas:
    ProductRequest:
      type: object
      properties:
        productName:
          type: string
          maxLength: 250
        description:
          type: string
        price:
          type: number
          minimum: 0
        active:
          type: boolean
          default: true
        taxRate:
          type: number
          minimum: 0
          default: 0
        discount:
          type: number
          minimum: 0
          maximum: 100
      required:
        - productName
        - price

    productResponse:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        productName:
          type: string
        description:
          type: string
        price:
          type: number
        active:
          type: boolean
        taxRate:
          type: number
        discount:
          type: number